refactor: move files to util-bootstrapping directory
