+++ packages/spectacular/src/lib/application-testing/util-bootstrapping/bootstrap-component-options.ts
